refactor(curation): robustez da CLI (paginação, slug em draft, atomicidade)
- queries: first:100 nas pendentes (evita truncar em 10 em silêncio); GROUP_BY_SLUG em stage:DRAFT (enxerga rascunhos não publicados); nova SUGGESTION_BY_ID_QUERY para o --id resolver no servidor.
- scripts: split em curate-runner.ts (lógica testável, import sem efeitos colaterais) + curate.ts (só bootstrap). applyCreate falha alto se o mark quebrar após criar o Group (evita duplicata na re-execução); contador 'erros' separado de 'pulados' no resumo.
- tests: 11 testes do driver com cliente mockado (slug único, applyCreate/applyUpdate, falha pós-create, múltiplas agendas, dryRunSummary, formatSuggestion).
- docs: nota de recuperação quando um CREATE falha no meio.
